+++ metadata.rb
@@ -7,14 +7,16 @@ long_description 'Installs/Configures lenovo_systemupdate'
 version '0.1.0'
 chef_version '>= 13.0'
 
+supports 'windows'
+
 # The `issues_url` points to the location where issues for this cookbook are
 # tracked.  A `View Issues` link will be displayed on this cookbook's page when
 # uploaded to a Supermarket.
 #
-# issues_url 'https://github.com/<insert_org_here>/lenovo_systemupdate/issues'
+issues_url 'https://github.com/hieunba/lenovo_systemupdate/issues'
 
 # The `source_url` points to the development repository for this cookbook.  A
 # `View Source` link will be displayed on this cookbook's page when uploaded to
 # a Supermarket.
 #
-# source_url 'https://github.com/<insert_org_here>/lenovo_systemupdate'
+source_url 'https://github.com/hieunba/lenovo_systemupdate'