+++ CHANGELOG.md
@@ -6,6 +6,4 @@ This file is used to list changes made in each version of the lenovo_systemupdat
 
 Initial release.
 
-- change 0
-- change 1
-
+- Installs Lenovo System Update 5.07.0084